[ART-36] Use a Store instead of the classic Repository implementation (#37)
### :pushpin: References
* **Issue:** [ART-36](https://github.com/Sefford/Artdrian/issues/ART-36)
Closes #36 
Closes #27 

### :tophat: What is the goal?

Phase out the old Repository polling architecture to a reactive one via
Stores.

### How is it being implemented?

The Repository pattern is fine as we normally use it; however for
Compose we're looking for a more Reactive implementation by using a
Store pattern in between.

This allows us to treat Wallpapers as a global state of the app, instead
of querying the API all the time.

For this we have enabled a Room Database to store the information from
the backend; although we'll be enabling Http
Caches further down the line. This is also an improvement mentioned in
#27 .

### QA Risks

We have deactivated some tests to accomodate the transition.

### How can it be tested?

The app should behave the same at this stage, although is not fully
functional after the latest API changes.