[ART-48] Refine Store Behavior (#49)
### :pushpin: References
Closes #48 

### :tophat: What is the goal?

Refine some behaviors concerning the behavior of the Store.

### How is it being implemented?

- Added some granularity on Errors to help diagnosing issues
- Ktor will not honor the cache if there is a connectivity error, so I
updated the WallpaperNetworkDataStore to default to the Cache in case
there is a connectivity error. This behavior was not behaving similarly
when using a CIO Engine and a Mock Engine, so I had to build some
classes around to be able to mock the neccessary behavior on tests.
- Improved responses on Local DataStore behavior, added tests because I
was missing those
- In case of errros I will hold the existing error until I receive both
responses, to avoid showing a Network error, and then updating it with
he Cache information
- Fixed a small bug concerning responses sources on Wallpaper Store that
was causing the store to enter in a persistence loop.