[ART-50] Register connectivity changes (#51)
### :pushpin: References
Closes #50

### :tophat: What is the goal?

Monitor connectivity for 

### How is it being implemented?

I used a Network Callback plus the StateStorage to create a Store where
I can subscribe for changes on the connectivity.

With this I am able to discern between WiFi/Ethernet/Mobile and analyze
connectivity speeds.

I will use this to prepare a pre-fetch of the Wallpapers later on